+++ algorithms/base.py
@@ -27,6 +27,7 @@ GROUP_MICRO = ("Microclimate", "microclimate")
 GROUP_STANDARDS = ("Plan Standards and QA", "standards")
 GROUP_REPORT = ("Reporting and Dashboard", "reporting")
 GROUP_OPTIMIZE = ("Optimization", "optimization")
+GROUP_EQUITY = ("Equity", "equity")
 
 
 class PlanXAlgorithm(QgsProcessingAlgorithm):